@@ -40,7 +40,7 @@ namespace GDStoSVG
     public class StructureRef : Element
     {
         public string? StructureName = null;
-        public Transform Transform = new Transform();
+        public Transform Transform = new();
 
         public override bool Check() => this.StructureName != null && this.Coords != null;
     }
@@ -48,7 +48,7 @@ namespace GDStoSVG
     public class ArrayRef : Element
     {
         public string? StructureName = null;
-        public Transform Transform = new Transform();
+        public Transform Transform = new();
         public Tuple<short, short>? RepeatCount;
 
         public override bool Check() => this.StructureName != null && this.Coords != null && this.RepeatCount != null;
@@ -58,7 +58,7 @@ namespace GDStoSVG
     {
         public short? Layer = null;
         public int Width = 0; // negative means not affected by magnification
-        public Transform Transform = new Transform();
+        public Transform Transform = new();
         public short? TextType = null;
         public byte Font = 0;
         public VerticalAlign VerticalPresentation = VerticalAlign.TOP;
@@ -97,9 +97,9 @@ namespace GDStoSVG
         public double Magnification = 1.0D;
         public double Angle = 0.0D; // degrees, counterclockwise
 
-        public Tuple<int, int> PositionOffset = new Tuple<int, int>(0, 0);
+        public Tuple<int, int> PositionOffset = new(0, 0);
 
-        public static readonly Transform Default = new Transform();
+        public static readonly Transform Default = new();
 
         /// <summary> Applies a transform from a parent element onto this one to produce a compound transform. </summary>
         /// <param name="trans"> The parent element's transform. </param>
